+++ .build/store-test.js
@@ -20,6 +20,7 @@ const BENIGN = [
   "Not implemented: Window's scrollTo",
   "Not implemented: window.scrollTo",
   "Not implemented: navigation",
+  "Not implemented: HTMLMediaElement",
 ];
 
 (async () => {